Add flag to disable sshd config

+++ linux-hardening/main/configure_sshd.sh
@@ -4,8 +4,11 @@ source ./remove_package.sh
 source ../first30.sh
 
 function configure_sshd() {
-    # Disable for now, configure later
-    systemctl stop sshd 
-    systemctl disable sshd
-    remove_package "$package_manager" "openssh-server"
+    flag="$1"
+    if [ "$flag" = "true" ]; then
+        # Disable for now, configure later
+        systemctl stop sshd 
+        systemctl disable sshd
+        remove_package "$package_manager" "openssh-server"
+    fi
 }
